Move editor.ts and window.ts to settings/
